+++ CCL.Creator/Validators/CarPackValidator.cs
@@ -1,0 +1,16 @@
+﻿using CCL.Types;
+using UnityEditor;
+
+namespace CCL.Creator.Validators
+{
+    internal class CarPackValidator : EditorWindow
+    {
+        public static void ValidateExport(CustomCarPack pack)
+        {
+            foreach (var item in pack.Cars)
+            {
+
+            }
+        }
+    }
+}
